server/requirements.txt
amqp==5.1.1
billiard==3.6.4.0
black==22.12.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Flask==2.2.2
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pytz==2022.7.1
requests==2.28.2
six==1.16.0
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.2
